requirements.txt
asposestorage==1.0.2
einops==0.8.0
fvcore==0.1.5.post20221221
gradio==4.37.1
infinibatch==0.1.1
kornia==0.7.2
matplotlib==3.9.0
nltk==3.8.1
numpy
omegaconf==2.3.0
Pillow==10.3.0
pyarrow==16.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
PyYAML==6.0.1
scipy==1.14.0
setuptools==69.5.1
scikit-image
tabulate==0.9.0
timm==1.0.7
torch==2.3.1
torchvision==0.18.1
transformers==4.42.2
wandb==0.17.3
webunit==1.3.10
whisper==1.1.10
opencv-python
